task07_smart_story/requirements.txt
# Task 7 - Smart Story
python-dotenv>=1.0.0
together>=0.2.0
requests>=2.28.0
